moved from rust experimental benchmark functionality to more stable `criterion` crate
